Safari-609.1.11.1/LayoutTests/http/tests/referrer-policy-script/origin/cross-origin-http-http.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of origin referrer policy when cross origin.");
jsTestIsAsync = true;

function checkReferrer(value) {
    referrer = value;
    // Should be the origin
    shouldBeEqualToString("referrer", "http://127.0.0.1:8000/");
    finishJSTest();
}

if (window.testRunner) {
    testRunner.setStatisticsShouldDowngradeReferrer(false, function () { downgradeReferrerCallback('origin', 'localhost')});
}
</script>
</body>
</html>
